+++ DSCResources/IisSite/IisSite.schema.psm1
@@ -112,10 +112,12 @@ Configuration IisSite
     . $userSettingsPath
     ##### END DO NOT MODIFY #####
 
-    $technology = [Technology]::Windows
+    $technology        = [Technology]::Windows
     $technologyVersion = [TechnologyVersion]::New( $OsVersion, $technology )
-    $technologyRole = [TechnologyRole]::New( "IISSite", $technologyVersion )
-    $stigDataObject = [STIG]::New( $StigVersion, $orgSettingsObject, $technology, $technologyRole, $technologyVersion, $exceptionsObject , $skipRuleTypeObject, $skipRuleObject )
+    $technologyRole    = [TechnologyRole]::New( "IISSite", $technologyVersion )
+    $stigDataObject    = [STIG]::New( $StigVersion, $OrgSettings, $technology,
+                                          $technologyRole, $technologyVersion, $Exception,
+                                          $SkipRuleType, $SkipRule )
 
     #### BEGIN DO NOT MODIFY ####
     # $StigData is used in the resources that are dot sourced below