+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added the following STIGs
+  * Windows Server 2016 V1R6
+
 ## 2.3.0.0
 
 * Windows 10 Fixes