+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Versions
 
 ## Unreleased
+* Fixed Issue [#207](https://github.com/Microsoft/PowerStig/issues/207): SQL Server Database rules fail to apply
+* Fixed Issue [#208](https://github.com/Microsoft/PowerStig/issues/208): Update PowerSTIG to use SQLServerDsc 12.1.0.0
 
 ## 2.3.0.0